+++ lib/generators/graphql/core.rb
@@ -4,6 +4,16 @@ require 'rails/generators/base'
 module Graphql
   module Generators
     module Core
+      def self.included(base)
+        base.send(
+          :class_option,
+          :directory,
+          type: :string,
+          default: "app/graphql",
+          desc: "Directory where generated files should be saved"
+        )
+      end
+
       def insert_root_type(type, name)
         log :add_root_type, type
         sentinel = /GraphQL\:\:Schema\.define do\s*\n/m
@@ -14,13 +24,13 @@ module Graphql
       end
 
       def create_mutation_root_type
-        create_dir("app/graphql/mutations")
-        template("mutation_type.erb", "app/graphql/types/mutation_type.rb", { skip: true })
+        create_dir("#{options[:directory]}/mutations")
+        template("mutation_type.erb", "#{options[:directory]}/types/mutation_type.rb", { skip: true })
         insert_root_type('mutation', 'MutationType')
       end
 
       def schema_file_path
-        "app/graphql/#{schema_name.underscore}.rb"
+        "#{options[:directory]}/#{schema_name.underscore}.rb"
       end
 
       def create_dir(dir)