+++ lib/graphql/backtrace/table.rb
@@ -89,12 +89,12 @@ module GraphQL
             rows << @cached_entries[id]
           else
             field_alias = ctx.ast_node.alias
-            row = [
+              row = [
               "#{position}",
               "#{field_name}#{field_alias ? " as #{field_alias}" : ""}",
               "#{ctx.object.inspect}",
               ctx.irep_node.arguments.to_h.inspect,
-              Backtrace::InspectResult.inspect(top && @override_value ? @override_value : ctx.value),
+              Backtrace::InspectResult.inspect_result(top && @override_value ? @override_value : ctx.value),
             ]
             rows << row
             @cached_entries[id] = row
@@ -116,7 +116,7 @@ module GraphQL
             "#{op_type}#{op_name ? " #{op_name}" : ""}",
             "#{query.root_value.inspect}",
             query.variables.to_h.inspect,
-            Backtrace::InspectResult.inspect(query.context.value),
+            Backtrace::InspectResult.inspect_result(query.context.value),
           ]
         else
           raise "Unexpected get_rows subject #{context_entry.inspect}"