+++ javascript_client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # graphql-ruby-client
 
+## 1.1.3 (11 Oct 2017)
+
+- Fix Apollo + ActionCable unsubscribe function #1019
+
+## 1.1.2 (9 Oct 2017)
+
+- Add channel IDs to ActionCable subscriptions #1004
+
 ## 1.1.1 (21 Sept 2017)
 
 - Add `--add-typename` option to `sync` #967